refactor: introduce `Workflow` type to easily build workflows
This commit introduces a `Workflow` type that can be leveraged to build
workflows more easily & in a more standardised fashion. to go with that,
this commit also introduces a `Flag` type to easily define & re-use (and
potentially access) defined flags of commands.

To demonstrate this new functionality, this commit also changes the
`local_workflows` package to use it.

While doing so, I've also refactored the implementations a bit.