chore: enable golangci-lint
Adding golangci-lint to the lint Makefile target, which means it will gate CI.

This PR is a little long, but I've tried to scope it down to a handful of
rulesets whose changes are mechanical and generally low-risk (do not introduce
significant logic changes, breaking API changes or refactoring).

I've also commented out the rules that I'd like to revisit in followups,
especially the ones that will require careful review and possibly some
discussion to address.