feat: clear auth config keys when switching auth types
Authenticating with token (deprecated) or oauth causes different config keys to
be written to the config store. If these are unset, it potentially leaves stale
auth information in the config. This also makes it difficult to determine which
auth information is being used during CLI interactions.

This change forces the authentication config to be either oauth or token
(deprecated). If one is set, the other method's keys are unset.

This required the addition of a Configuration.Unset method. "Unset" is a
long-sought after feature of spf13/viper, which will probably never be added by
the maintainer, but it is fairly straightforward to implement in coordination
with the config file storage.