chore: transform new fingerprints
chore: add new fingerprints to mock sarif for testing

chore(fingerprint): rename CodeSast types to Fingerprint types

The changes in this commit rename the `TypesCodeSastFingerprintProjectV1`, `TypesCodeSastFingerprintRepositoryV1`, and `TypesCodeSastFingerprintAssetV1` types to `TypesFingerprintProjectV1`, `TypesFingerprintRepositoryV1`, and `TypesFingerprintAssetV1` respectively. This is to align the naming convention with the broader Fingerprint types and make the codebase more consistent.

chore(data-transformation): Improve fingerprint parsing and validation in tests