chore: use `go tool`
Stop using global `mockgen` and `oapi-codegen`, now use the `go tool` command to have a project specified version of `mockgen` and `oapi-codegen`.
With the new tool approach in `go.mod`, `tools.go` was no longer needed and has been removed.
Use to oapi-codegen v2.4.1, to be aligned with the version in go.mod. Regenerated all files after the update.